==> 01.ETL_steam_games.ipynb <==
ETL Steam Games

#Importando librerias
import json
import numpy as np
import pandas as pd
import ast

1. Extraccion de datos

steam_1 = pd.read_json('data/steam_games.json.gz',lines=True,compression='gzip')
steam_1.head(3)

2. Exploración de datos

#reviso si tiene algun dato
steam_1.tail(3)

2.1 Busco valores nulos

steam_1.isna().sum() #Buscamos y sumamos los valores nulos por cada categoria.

#Porcentajes de ratios de nulos
na_ratio = ((steam_1.isnull().sum() / len(steam_1))*100).sort_values(ascending = False)
print(na_ratio)

#veo cuantos registros se eliminaran
steam_1.shape

#Elimino todos los nulos
steam_2 = steam_1.dropna(how="all").reset_index(drop=True)
steam_2.shape

2.2 Busco valores repetidos

steam_3=steam_2.drop_duplicates(subset="id")
steam_3.shape

2.3 Algunas variables parecen ser listas anidadas, esto es un formato usual cuando se guarda en formato jason,
    por lo cual se revisaran cuantas variables se encuentran asi y desanidarlas en el caso que se necesite usarlas.

tipo_data = {"columna":[],"tipos_de_datos":[]} #genero un diccionario vacio para ir almacenando lo que genere el bucle

for columna in steam_3.columns: #un bucle que va recorriendo 
    tipo_data["columna"].append(columna)
    tipo_data["tipos_de_datos"].append(steam_3[columna].apply(type).unique())

analisis= pd.DataFrame(tipo_data)
analisis

Todos aquellos que tengan 'list', significa que hay una lista anidada dentro de esa variable

Para crear las funciones solamente requerimos de 4 variables de la tabla steam, pero luego para calcular el modelo de recomedanción se requiere realizar un EDA para verificar cuales serian funcionales al modelo. Por lo cual en una primera instancia transformaremos algunas variables y luego crearemos dos tablas, una exclusivamente para las 4 funciones y otra para el modelo de recomendación.
Esto se realiza debido al consumo de memoria de la API es limitada, lo cual para ser más eficiente, se requiere las bases de datos exclusivamente las necesarias.

2.4 Creación de año

steam_3['Año de lanzamiento'] = steam_3['release_date'].str.extract(r'(\d{4})').fillna('0')
steam_3.head(3)

2.5 Transformación de id, para poder realizar cruces

#Para convertir ID, debo asegurarme que no haya NaN
print(steam_3[steam_3['id'].isnull()])

steam_4 = steam_3.dropna(subset=['id'])
steam_4.shape

#transformo y renombro
steam_4['id']=steam_4['id'].apply(np.int64)
steam_4= steam_4.rename(columns={"id":"item_id"})

#verifico
steam_4.dtypes

2.6 Transformar la variable genero

Se crean las filas necesarias por la cantidad de generos registrados por cada id, con lo cual si un id tiene 5 generos, ahora habrá 5 ids repetidos pero con generos unicos.

steam_5 = steam_4.explode("genres")
steam_5.shape

steam_5.head(3)

2.7 Me quedo con las variables que son necesarias

 steam_6= steam_5[['item_id','app_name','title','genres','Año de lanzamiento','publisher','developer','specs','price']].copy()
steam_6.head(2) #queda como la tabla final para el sistema de recomendación

steam_7 =  steam_6[['item_id','app_name','title','genres','Año de lanzamiento']].copy()
steam_7.head(2) #queda como la tabla final para las funciones

2.8 guardo las tablas para reutilizarlas más adelante

steam_7_csv = "data/steam_7_api.csv"
steam_7.to_csv(steam_7_csv , index=False, encoding="utf-8")

steam_6_csv = "data/steam_6_modelo.csv"
steam_6.to_csv(steam_6_csv , index=False, encoding="utf-8")

*****************************************************
