==> 05.funciones.ipynb <==
## Desarrollo de funciones para la Api

+ def **PlayTimeGenre( *`genero` : str* )**:
    Debe devolver `año` con mas horas jugadas para dicho género.
  
    Ejemplo de retorno: {"Año de lanzamiento con más horas jugadas para Género X" : 2013}

+ def **UserForGenre( *`genero` : str* )**:
    Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año.

    Ejemplo de retorno: {"Usuario con más horas jugadas para Género X" : us213ndjss09sdf,
			     "Horas jugadas":[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas: 23}]}

+ def **UsersRecommend( *`año` : int* )**:
    Devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = True y comentarios positivos/neutrales)
  
    Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

+ def **UsersNotRecommend( *`año` : int* )**:
   Devuelve el top 3 de juegos MENOS recomendados por usuarios para el año dado. (reviews.recommend = False y comentarios negativos)
  
   Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

+ def **sentiment_analysis( *`año` : int* )**:
    Según el año de lanzamiento, se devuelve una lista con la cantidad de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento. 

    Ejemplo de retorno: {Negative = 182, Neutral = 120, Positive = 278}


#Importar librerias
import pandas as pd
import numpy as np 
import pyarrow as pa
import pyarrow.parquet as pq

### 1. Función **def PlayTimeGenre**(genero:str):


Requiero de las tablas steam_games que contiene el genero y año de lanzamiento. Por otro lado la tabla
user_items que contiene la variable playtimeforever, que sería horas/minutos judados

#Traigo la tabla steam games
steam_games= pd.read_csv('data/steam_7_api.csv', encoding='utf8')
steam_games.head(2)

#Se extraen las columnas para generar un nuevo dataframe y correr la función
genre= steam_games[["item_id","genres","Año de lanzamiento"]]

#Traigo la tabla user_items para unirlas
user_items= pd.read_csv('data/items_3.csv', encoding='utf8')
user_items.head(2) 

#me quedo solo con la variable que necesito
play_time= user_items[["item_id","playtime_forever"]]
play_time.head(3)

#unifico
genre_2= play_time.merge(genre, on="item_id")
genre_2.head(5)

Entonces ahora debo agrupar por genero y año las horas jugadas
+ Año de lanzamiento con más horas jugadas para Género X

horas_anio_genre=genre_2.groupby(["genres","Año de lanzamiento"]).playtime_forever.sum().reset_index(name='Horas jugadas')
horas_anio_genre.sort_values(by="Horas jugadas", ascending=False)

+ Año de lanzamiento con más horas jugadas para Género X, ademas el formato string para genero

horas_anio_genre['genres']=horas_anio_genre['genres'].astype('string') # pasar de objeto a string
horas_anio_genre['Año de lanzamiento']=horas_anio_genre['Año de lanzamiento'].astype('string')

 debemos encontrar el maximo de horas jugadas por genero

res = horas_anio_genre.groupby("genres")["Horas jugadas"].agg(Horas="max").reset_index()
res.head()


res.shape

#para realizar el join, todas las variables de deben llamar iguales, asi cuando haga el merge, me buscara solamente la info que
#necesito para completar mi tabla res, que contiene las maximas horas por genero
genero2 = horas_anio_genre.rename(columns={'Horas jugadas':'Horas'})
genero2


tabla_final =pd.merge(res,genero2)
tabla_final

def PlayTimeGenre(genero):
    anio= tabla_final[tabla_final["genres"]==  genero ]["Año de lanzamiento"].iloc[0]
    horas_jugadas= tabla_final[tabla_final["genres"]== genero]["Horas"].iloc[0]
    return {"El género": genero, "en el año de lanzamiento": int(anio),"tiene más horas jugadas": int(horas_jugadas)}

PlayTimeGenre('Action')

#se guarda para la api
tabla_final2 = "funcion_1.csv"
tabla_final.to_csv(tabla_final2, index=False, encoding="utf-8")

#convertir a parquet las funciones

#Transformo el archivo csv a parquet
#Leo el archivo csv
funcion_1= pd.read_csv("funcion_1.csv") 

#Indico donde quiero guardar el parquet y con que nombre
output_file= "data/funcion_1.parquet"

#Transformo a traves de una tabla el archivo csv en parquet
table = pa.Table.from_pandas(funcion_1)
pq.write_table(table,output_file)

### 2. Función **def UserForGenre(genero:str)**:

Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año.

    Ejemplo de retorno: {"Usuario con más horas jugadas para Género X" : us213ndjss09sdf,
			     "Horas jugadas":[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas: 23}]}

#Sigo con la misma tabla de steams + items, pero ahora incluyo al user_id
#el user_id esta como object, lo tengo que pasar a string, asi que medio que debo a rehacer casi todo lo de la funcion 1
user_items.head(2)

user_items2=user_items.copy() #creo una tabla extra por las dudas

user_items2['user_id']=user_items2['user_id'].astype('string') #paso de object a string

play_time_user= user_items[["item_id","playtime_forever","user_id"]] # me quedo con las variables necesarias
play_time_user.head(3)

genre.head(3) #verifico antes mi tabla a curzar

#unifico
genre_user= play_time_user.merge(genre, on="item_id")
genre_user.head(5)

genre_user['genres']=genre_user['genres'].astype('string') #pasar de object a string
genre_user['Año de lanzamiento']=genre_user['Año de lanzamiento'].astype('string')
genre_user['user_id']=genre_user['user_id'].astype('string')

#agrupo el genero por usuario y horas jugadas
user_hours_year= genre_user.groupby(["genres","user_id","Año de lanzamiento"])["playtime_forever"].sum().reset_index()
user_hours_year

#necesito saber los usuarios con mayor horas jugadas por genero
user_hours_total= genre_user.groupby(["genres","user_id"])["playtime_forever"].sum().reset_index()
user_hours_total.head(3)

#por cada genero el total de horas sumadas
max_user = user_hours_total.groupby(["genres"])["playtime_forever"].agg(playtime_forever="max").reset_index()
max_user

tabla_user =pd.merge(max_user,user_hours_total)
tabla_user #ya tengo el usuario con mayor horas jugadas por genero, ahora me falta incluir el total por año

user_hours_year2=user_hours_year.rename(columns={'playtime_forever':'playtime_forever_year'})
user_hours_year2.head(5)

#voy a probar si el cruce funciona ok, en el caso del usuario REBAS_AS_F-T	
prueba3=user_hours_year2[(user_hours_year2['user_id'] == 'REBAS_AS_F-T') & (user_hours_year2['genres']=='Indie')]
prueba3
#Para el usuario REBAS_AS_F-T la tabla tabla_user me debe traer 16 registros

tabla_user2 =pd.merge(tabla_user,user_hours_year2)
tabla_user2

#verificamos
prueba4=tabla_user2[(tabla_user2['user_id'] == 'REBAS_AS_F-T') & (tabla_user2['genres']=='Indie')]
prueba4.shape # esta ok
#ya tenemos la tabla para crear la funcion

tabla_user2 = tabla_user2.rename(columns={'Año de lanzamiento':'Año', 'playtime_forever_year':'Horas jugadas'})
tabla_user2["Año"] = pd.to_numeric(tabla_user2["Año"])

def UserForGenre(genero):
    usuario= tabla_user2[tabla_user2["genres"]== genero]["user_id"].iloc[0] #obtengo usuario
    historial=tabla_user2[(tabla_user2['user_id'] == usuario) & (tabla_user2['genres']==genero)] #filtro por el genero y usuario
    historial2 = historial[['Año', 'Horas jugadas']].copy() #me quedo con las columnas necesarias
    historial3=historial2.to_dict(orient="records")
    return {"Usuario":usuario ,"con más horas jugadas para": genero, "Historial acumulado": historial3 }

UserForGenre('Indie')

#se guarda para la api
tabla_user3 = "funcion_2.csv"
tabla_user2.to_csv(tabla_user3, index=False, encoding="utf-8")

#guardo parquet
funcion_2= pd.read_csv("funcion_2.csv") 

#Indico donde quiero guardar el parquet y con que nombre
output_file= "data/funcion_2.parquet"

#Transformo a traves de una tabla el archivo csv en parquet
table = pa.Table.from_pandas(funcion_2)
pq.write_table(table,output_file)

### 3. Funcion **defUsersRecommend(año:int)**:

Devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = True y comentarios positivos/neutrales)
  
    Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

Solo uso la tabla de reviews, ya que seria la recomendación del año dado, y la misma contiene posted, el momento que se publico
Utilizar el año de lanzamiento mezclando con comentarios de otro año, podria no ser consistente.
De todas maneras steams, te dice el nombre

games_users= steam_games[["item_id","app_name","Año de lanzamiento"]]

#saco los duplicados por genero, ya que en este caso no lo utilizaremos
games_users=game_users.drop_duplicates()

games_users.shape

games_users.dtypes

reviews= pd.read_csv('data/reviews_9.csv', encoding='utf8')
reviews.head(2)

reviews2= reviews[["user_id","item_id","recommend","year"]] # me quedo solo con las variables necesarias

#unifico
tabla_reviews= games_users.merge(reviews2, on="item_id", how="right")
tabla_reviews.shape

tabla_reviews=tabla_reviews.drop_duplicates()
tabla_reviews.shape #vs 59305, ahora hay menos casos

#convierto la tabla recommend de boleando a 1/0
tabla_reviews["recommend"] = tabla_reviews["recommend"].astype(int)
tabla_reviews.head(5)

top_reviews=tabla_reviews.groupby(["item_id","app_name","year"]).recommend.sum().reset_index()
top_reviews

max_reviews = top_reviews.groupby(["year"])["recommend"].nlargest(3).reset_index()
max_reviews

max_reviews2=pd.merge(max_reviews,top_reviews)
max_reviews2

#ahora necesito ordenarlo en el top 3, y agregar puesto 1, 2, 3 
max_reviews2.sort_values(by = ['year', 'recommend'], ascending=False)
max_reviews2['Acumulado'] = max_reviews2.groupby([max_reviews2['year']])['year'].cumcount()
max_reviews2['Acumulado']=max_reviews2['Acumulado'].replace([0,1,2],['Puesto 1','Puesto 2','Puesto 3'])
max_reviews3=max_reviews2.drop(columns=['level_1'])
max_reviews3

#volvemos a string el año , porque la api lo toma como string y no me deja poner int
max_reviews3['year']=max_reviews3['year'].apply(str)

def UsersRecommend(anio:int): #el int hay que poner en la api para que funcione, este codigo corrio
    tabla1=max_reviews3[max_reviews3['year'] == anio]
    tabla1.reset_index()
    
    dato = tabla1[tabla1["year"]== anio]["app_name"].iloc[0]
    dato1 = tabla1[tabla1["year"]== anio]["app_name"].iloc[1]
    dato2 = tabla1[tabla1["year"]== anio]["app_name"].iloc[2]
    
    return {"Los juegos más recomendados para el año": año, "Puesto 1": dato,"Puesto 2": dato1,"Puesto 3": dato2}

def UsersRecommend(anio): #este  codigo no corrio en la api
    anio=str(anio)
    tabla1=max_reviews3[max_reviews3['year'] == anio]
    tabla1.reset_index()
    
    dato = tabla1[tabla1["year"]== anio]["app_name"].iloc[0]
    dato1 = tabla1[tabla1["year"]== anio]["app_name"].iloc[1]
    dato2 = tabla1[tabla1["year"]== anio]["app_name"].iloc[2]
    
    return {"Los juegos más recomendados para el año": anio, "Puesto 1": dato,"Puesto 2": dato1,"Puesto 3": dato2}

UsersRecommend(2014)

#se guarda para la api
max_reviews4 = "funcion_3.csv"
max_reviews3.to_csv(max_reviews4, index=False, encoding="utf-8")

funcion_3= pd.read_csv("funcion_3.csv") 

#Indico donde quiero guardar el parquet y con que nombre
output_file= "data/funcion_3.parquet"

#Transformo a traves de una tabla el archivo csv en parquet
table = pa.Table.from_pandas(funcion_3)
pq.write_table(table,output_file)

### 4. Funcion **def UsersNotRecommend( año:int )**:

Devuelve el top 3 de juegos MENOS recomendados por usuarios para el año dado. (reviews.recommend = False y comentarios negativos)

    Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

#Parto de la funcion anterior, en vez del max, el min
min_reviews = top_reviews.groupby(["year"])["recommend"].nsmallest(3).reset_index()
min_reviews.head(2)

min_reviews2=pd.merge(min_reviews,top_reviews)
min_reviews2.head(5) #hay muchos juegos sin recomendaciones, pero elegimos los primeros 3 que aparezcan en la tabla

min_reviews2.sort_values(by = ['year', 'recommend'], ascending=False)
min_reviews2=min_reviews2.drop(columns=['level_1'])
min_reviews2['year']=min_reviews2['year'].astype('string')

def UsersNotRecommend(año):
    año=str(año)
    tabla11=min_reviews2[min_reviews2['year']==año]
    tabla22=tabla11[['app_name','year']].copy() #me quedo con las columnas necesarias
    tabla22.reset_index()
    
    dato0 = tabla22[tabla22["year"]==año]["app_name"].iloc[0]
    dato11 = tabla22[tabla22["year"]== año]["app_name"].iloc[1]
    dato22 = tabla22[tabla22["year"]==año]["app_name"].iloc[2]
    
    return {"Los juegos menos recomendados para el año": año, "Puesto 1": dato0,"Puesto 2": dato11,"Puesto 3": dato22}

UsersNotRecommend(2014)

#se guarda para la api
min_reviews3 = "funcion_4.csv"
min_reviews2.to_csv(min_reviews3, index=False, encoding="utf-8")

funcion_4= pd.read_csv("funcion_4.csv") 

#Indico donde quiero guardar el parquet y con que nombre
output_file= "data/funcion_4.parquet"

#Transformo a traves de una tabla el archivo csv en parquet
table = pa.Table.from_pandas(funcion_4)
pq.write_table(table,output_file)

### 5. Función **def sentiment_analysis(año:int)**:


Según el año de lanzamiento, se devuelve una lista con la cantidad de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento. 

    Ejemplo de retorno: {Negative = 182, Neutral = 120, Positive = 278}

games_users.head() #reviso la tabla a utilizar

games_users=game_users.drop_duplicates()

games_users.shape

reviews.head(2)

sentimiento_analysis=reviews.merge(games_users, on="item_id")
sentimiento_analysis.head(3)

sentimiento_analysis.shape

def sentiment_analysis(anio):
    
    reviews_por_anio=sentimiento_analysis[sentimiento_analysis["Año de lanzamiento"]== anio]
    
 
    Negativos = 0
    Neutral = 0
    Positivos = 0
    

    for i in reviews_por_anio["sentiment_analisis"]:
        if i == 0:
            Negativos += 1
        elif i == 1:
            Neutral += 1 
        elif i == 2:
            Positivos += 1

    count_sentiment ={"Negative": Negativos , "Neutral" : Neutral, "Positive": Positivos}
    
    return count_sentiment

sentiment_analysis(2015)

sentimiento_analysis2 = "funcion_5.csv"
sentimiento_analysis.to_csv(sentimiento_analysis2, index=False, encoding="utf-8")

funcion_5= pd.read_csv("funcion_5.csv") 

#Indico donde quiero guardar el parquet y con que nombre
output_file= "data/funcion_5.parquet"

#Transformo a traves de una tabla el archivo csv en parquet
table = pa.Table.from_pandas(funcion_5)
pq.write_table(table,output_file)
